Add IterableTrait to implement Iterator interface
