+++ src/php/Lang/IterableTrait.php
@@ -1,0 +1,31 @@
+<?php
+
+namespace Phel\Lang;
+
+trait IterableTrait
+{
+    public function current()
+    {
+        return current($this->data);
+    }
+
+    public function key()
+    {
+        return key($this->data);
+    }
+
+    public function next(): void
+    {
+        next($this->data);
+    }
+
+    public function rewind(): void
+    {
+        reset($this->data);
+    }
+
+    public function valid(): bool
+    {
+        return key($this->data) !== null;
+    }
+}
